+++ sync_diff_inspector/pkg/config/config_test.go
