+++ sync_diff_inspector/pkg/config/template.go
