+++ sync_diff_inspector/pkg/progress/progress_test.go
