+++ sync_diff_inspector/pkg/report/report_test.go
@@ -24,9 +24,9 @@ import (
 	"github.com/BurntSushi/toml"
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/pingcap/tidb-tools/pkg/dbutil"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/chunk"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/config"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/source/common"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/chunk"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/config"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/source/common"
 	"github.com/pingcap/tidb/parser"
 	"github.com/stretchr/testify/require"
 )
