+++ sync_diff_inspector/pkg/source/common/table_diff.go
