+++ sync_diff_inspector/pkg/source/source_test.go
@@ -27,11 +27,11 @@ import (
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/pingcap/tidb-tools/pkg/dbutil"
 	filter "github.com/pingcap/tidb-tools/pkg/table-filter"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/chunk"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/config"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/source/common"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/splitter"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/utils"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/chunk"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/config"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/source/common"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/splitter"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/utils"
 	"github.com/pingcap/tidb/parser"
 	router "github.com/pingcap/tidb/util/table-router"
 	"github.com/stretchr/testify/require"
