+++ sync_diff_inspector/pkg/source/tidb.go
@@ -25,10 +25,10 @@ import (
 	"github.com/pingcap/log"
 	"github.com/pingcap/tidb-tools/pkg/dbutil"
 	"github.com/pingcap/tidb-tools/pkg/filter"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/config"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/source/common"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/splitter"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/utils"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/config"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/source/common"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/splitter"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/utils"
 	"github.com/pingcap/tidb/parser/model"
 	"go.uber.org/zap"
 )
