+++ sync_diff_inspector/pkg/splitter/index_fields_test.go
