+++ sync_diff_inspector/pkg/splitter/splitter_test.go
@@ -23,9 +23,9 @@ import (
 
 	sqlmock "github.com/DATA-DOG/go-sqlmock"
 	"github.com/pingcap/tidb-tools/pkg/dbutil"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/chunk"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/source/common"
-	"github.com/pingcap/tidb-tools/sync_diff_inspector/utils"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/chunk"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/source/common"
+	"github.com/pingcap/tidb-tools/sync_diff_inspector/pkg/utils"
 	"github.com/pingcap/tidb/parser"
 	"github.com/stretchr/testify/require"
 )
